Refactoring: use event delegation
