+++ public/index.html
@@ -17,9 +17,14 @@
         </span>
       </div>
       <div class="category">
-        <button class="category-btn">
+        <button class="category-btn layer-btn">
           <img class="category-img" src="/img/categoryBtn.png">카테고리
         </button>
+        <div class="category-layer non-display">
+          <div class="layer-container"></div>
+          <div class="layer-container"></div>
+          <div class="layer-container"></div>
+        </div>
         <span>|</span>
         <button class="category-btn">핫딜</button>
         <button class="category-btn">베스트100</button>
